Create hello world using x64
Here is the description for what all the function does in the given program

mov rax, 1: This instruction copies the value 1 into the rax register. The value 1 is the system call number for write, which writes data to a file descriptor.

mov rdi, 1: This instruction copies the value 1 into the rdi register. The value 1 is the file descriptor for stdout.

mov rsi, msg: This instruction copies the address of the string msg into the rsi register. rsi is used to pass the pointer to the data to be written.

mov rdx, 13: This instruction copies the value 13 into the rdx register. rdx is used to pass the length of the data to be written.

syscall: This instruction triggers a system call, which is a signal to the processor to stop what it’s doing and do something else. The system call number and arguments are taken from the rax, rdi, rsi, and rdx registers.

mov rax, 60: This instruction copies the value 60 into the rax register. The value 60 is the system call number for exit, which terminates the program.

mov rdi, 0: This instruction copies the value 0 into the rdi register. rdi is used to pass the exit status to the exit system call.

syscall: This triggers the exit system call, which terminates the program